# JavaDoc Infrastructure
* Started documenting based on JavaDoc convensions
* Created an initial example of JavaDoc + JavaDoc generation settings.

# Client-Side Improvements
* Completely reorganized file structure for better readability
* Separated the "connect" and "login" methods in the login form.
* Made all fxml addresses a global constant.
* Created a custom Runnable interface called `RunnableWithMessage` which supports passing a message to a regular Runnable object - this allows handling messages in preset runnables and allows better code centralization.
* Created a custom Task class called 'MessageTasker' to handle client-side message handling. It handles all the queueing, sending, receiving and parsing of data, and will also update the waiting screen interface.
To use it, you need to create a new instance of `MessageTasker` and supply at least the Message object to sent to the server, and two `RunnableWithMessage` objects - one for success and one for failure. You can also optionally supply success and failure strings, otherwise it uses defaults.

#GUI Improvements
* "Enter" key in login screen will now attempt login
* A countdown will now appear on the "OK" button when a waiting window is slated to self-close.
* Closing the window will now prompt user if he's sure and actually quit the program (it just hid the window before).
* Waiting screen now also disables all other elements to prevent key presses.

# Client/Server Implementation
* Updated `MessageHandler` on Server Side to work with the client on creating new PreOrders.  Currently it creates a dummy order.
* Changed UserID to `Integer` rather than `long`
* Open orders and existing parking lots are now populated by the server in real time!

+++ src/main/entity/Customer.java
@@ -1,8 +1,6 @@
 package entity;
 
 import java.util.ArrayList;
-import entity.Order;
-import entity.Subscription;
 
 
 /**
@@ -25,7 +23,7 @@ public class Customer extends User{
 	 * @param carIDList Customer's cars' ids list.
 	 * @param subscriptionList Customer's subscriptions' list.
 	 */
-	public Customer(long uID, String name, String email, ArrayList<Integer> carIDList, ArrayList<Subscription> subscriptionList) {
+	public Customer(Integer uID, String name, String email, ArrayList<Integer> carIDList, ArrayList<Subscription> subscriptionList) {
 		super(uID,name,email);
 		this._carIDList = carIDList;
 		this._subscriptionList = new ArrayList<>();
